refactor: remove Cache trait and module
Move SimpleCache as OfflineSolver to solver.rs.
